+++ back_local/src/main/java/com/portfolioweb/aj/Repositorio/RTipoEmpleo.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RTipoEmpleo extends JpaRepository<TipoEmpleo, Long> {
+public interface RTipoEmpleo extends JpaRepository<TipoEmpleo, Integer> {
 
     Optional<TipoEmpleo> findByNombre(String nombre);
