+++ back_local/src/main/java/com/portfolioweb/aj/Repositorio/RTipoHabilidad.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RTipoHabilidad extends JpaRepository<TipoHabilidad, Long> {
+public interface RTipoHabilidad extends JpaRepository<TipoHabilidad, Integer> {
 }
